merge

+++ bin/node.sh
@@ -7,10 +7,10 @@ if [[ "$1" == "gateway" ]]; then
 fi
 
 if [[ "$1" == "agent" ]]; then
-  mkdir -p /var/run/pibox
-  FB_VERSION=v6
+  mkdir -p /var/run/pibox &> /dev/null
+  FB_VERSION="v6"
   FB_PATH=/var/run/pibox/pibox-framebuffer-$FB_VERSION
-  if [[ ! -f $FB_PATH ]]; then
+  if [[ ! -f $FB_PATH && -d /var/run/pibox ]]; then
     curl --connect-timeout 10 -sLo $FB_PATH https://github.com/kubesail/pibox-framebuffer/releases/download/$FB_VERSION/pibox-framebuffer
     chmod +x $FB_PATH
     rm -f /var/run/pibox/pibox-framebuffer
